// src/app/admin/pages/customers/interfaces/AdminCustomersInterface.ts
export interface IAdminCustomerInputs {
    customername: string
    contactname: string
    personaddress: string
    emailaddress: string
    phone1: string
    phone2: string
    taxno: string
    cno: string
    maxcredit: number
    maxtime: number
    intbalance: number
    accountno: string
}